% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/User_Codes.R
\name{BF_lambda_obs_LLOG}
\alias{BF_lambda_obs_LLOG}
\title{Outlier detection for observation for the log-logisitc model}
\usage{
BF_lambda_obs_LLOG(ref, obs, X, chain)
}
\arguments{
\item{ref}{Reference value \eqn{\lambda_{ref}} or \eqn{u_{ref}}}

\item{obs}{Indicates the number of the observation under analysis}

\item{X}{Design matrix with dimensions \eqn{n} x  \eqn{k} where \eqn{n} is
the number of observations and \eqn{k} is the number of covariates
(including the intercept).}

\item{chain}{MCMC chains generated by a BASSLINE MCMC function}
}
\description{
This returns a unique number corresponding to the Bayes Factor
associated to the test \eqn{M_0: \Lambda_{obs} = \lambda_{ref}} versus
\eqn{M_1: \Lambda_{obs}\neq \lambda_{ref}} (with all other
\eqn{\Lambda_j,\neq obs} free). The value of \eqn{\lambda_{ref}} is
required as input. The user should expect long running times for the
log-Student’s t model, in which case a reduced chain given
\eqn{\Lambda_{obs} = \lambda_{ref}} needs to be generated
}
\examples{
library(BASSLINE)

# Please note: N=1000 is not enough to reach convergence.
# This is only an illustration. Run longer chains for more accurate
# estimations.

LLOG <- MCMC_LLOG(N = 1000, thin = 20, burn = 40, Time = cancer[, 1],
                  Cens = cancer[, 2], X = cancer[, 3:11])
LLOG.Outlier <- BF_lambda_obs_LLOG(1,1, X = cancer[, 3:11], chain = LLOG)

}
