% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/User_Codes.R
\name{CaseDeletion_LEP}
\alias{CaseDeletion_LEP}
\title{Case deletion analysis for the log-exponential power model}
\usage{
CaseDeletion_LEP(Time, Cens, X, chain, set = TRUE, eps_l = 0.5, eps_r = 0.5)
}
\arguments{
\item{Time}{Vector containing the survival times.}

\item{Cens}{Censoring indication (1: observed, 0: right-censored).}

\item{X}{Design matrix with dimensions \eqn{n} x  \eqn{k} where \eqn{n} is
the number of observations and \eqn{k} is the number of covariates
(including the intercept).}

\item{chain}{MCMC chains generated by a BASSLINE MCMC function}

\item{set}{Indicator for the use of set observations (1: set observations,
0: point observations). The former is strongly recommended over the
latter as point observations cause problems in the context of Bayesian
inference (due to continuous sampling models assigning zero probability
to a point).}

\item{eps_l}{Lower imprecision \eqn{(\epsilon_l)} for set observations
(default value: 0.5).}

\item{eps_r}{Upper imprecision \eqn{(\epsilon_r)} for set observations
(default value: 0.5)}
}
\description{
Leave-one-out cross validation analysis. The function returns a
matrix with n rows. The first column contains the logarithm of the CPO
(Geisser and Eddy, 1979). Larger values of the CPO indicate better
predictive accuracy of the model. The second and third columns contain
the KL divergence between \eqn{\pi(\beta, \sigma^2,  \theta | t_{-i})}
and \eqn{\pi(\beta, \sigma^2,  \theta | t)} and its calibration index
\eqn{p_i}, respectively.
}
\examples{
library(BASSLINE)

# Please note: N=1000 is not enough to reach convergence.
# This is only an illustration. Run longer chains for more accurate
# estimations (especially for the log-exponential power model).

LEP <- MCMC_LEP(N = 1000, thin = 20, burn = 40, Time = cancer[, 1],
                Cens = cancer[, 2], X = cancer[, 3:11])
LEP.CD <- CaseDeletion_LEP(Time = cancer[, 1], Cens = cancer[, 2],
                           X = cancer[, 3:11], chain = LEP)

}
