% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/User_Codes.R
\name{CaseDeletion_LST}
\alias{CaseDeletion_LST}
\title{Case deletion analysis for the log-student's t model}
\usage{
CaseDeletion_LST(Time, Cens, X, chain, set, eps_l, eps_r)
}
\arguments{
\item{Time}{Vector containing the survival times.}

\item{Cens}{Censoring indication (1: observed, 0: right-censored).}

\item{X}{Design matrix with dimensions \eqn{n} x  \eqn{k} where \eqn{n} is
the number of observations and \eqn{k} is the number of covariates
(including the intercept).}

\item{chain}{MCMC chains generated by a BASSLINE MCMC function}

\item{set}{Indicator for the use of set observations (1: set observations,
0: point observations).}

\item{eps_l}{Lower imprecision \eqn{(\epsilon_l)} for set observations
(default value: 0.5).}

\item{eps_r}{Upper imprecision \eqn{(\epsilon_r)} for set observations
(default value: 0.5)}
}
\description{
Leave-one-out cross validation analysis. The function returns a
matrix with n rows. Its first column contains the logarithm of the CPO
(Geisser and Eddy, 1979). Larger values of the CPO indicate better
predictive accuracy of the model. The second and third columns contain
the KL divergence between \eqn{\pi(\beta, \sigma^2,  \theta | t_{-i})}
and \eqn{\pi(\beta, \sigma^2,  \theta | t)} and its calibration index
\eqn{p_i}, respectively.
}
