% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/User_Codes.R
\name{DIC_LLAP}
\alias{DIC_LLAP}
\title{Deviance information criterion for the log-Laplace model}
\usage{
DIC_LLAP(Time, Cens, X, chain, set = 1, eps_l = 0.5, eps_r = 0.5)
}
\arguments{
\item{Time}{Vector containing the survival times.}

\item{Cens}{Censoring indication (1: observed, 0: right-censored).}

\item{X}{Design matrix with dimensions \eqn{n} x  \eqn{k} where \eqn{n} is
the number of observations and \eqn{k} is the number of covariates
(including the intercept).}

\item{chain}{MCMC chains generated by a BASSLINE MCMC function}

\item{set}{Indicator for the use of set observations (1: set observations,
0: point observations). The former is strongly recommended over the
latter as point observations cause problems in the context of Bayesian
inference (due to continuous sampling models assigning zero probability
to a point).}

\item{eps_l}{Lower imprecision \eqn{(\epsilon_l)} for set observations
(default value: 0.5).}

\item{eps_r}{Upper imprecision \eqn{(\epsilon_r)} for set observations
(default value: 0.5)}
}
\description{
Deviance information criterion is based on the deviance funcion
\eqn{D(\theta, y) = -2 log(f(y|\theta))} but also incorporates a
penalization factor of the complexity of the model
}
\examples{
library(BASSLINE)

# Please note: N=1000 is not enough to reach convergence.
# This is only an illustration. Run longer chains for more accurate
# estimations.

LLAP <- MCMC_LLAP(N = 1000, thin = 20, burn = 40, Time = cancer[,1],
                  Cens = cancer[,2], X = cancer[,3:11])
LLAP.DIC <- DIC_LLAP(Time = cancer[,1], Cens = cancer[,2], X = cancer[,3:11],
                     chain = LLAP)

}
