% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/User_Codes.R
\name{DIC_LN}
\alias{DIC_LN}
\title{Deviance information criteria for the log-normal model}
\usage{
DIC_LN(Time, Cens, X, chain, set, eps_l, eps_r)
}
\arguments{
\item{Time}{Vector containing the survival times.}

\item{Cens}{Censoring indication (1: observed, 0: right-censored).}

\item{X}{Design matrix with dimensions \eqn{n} x  \eqn{k} where \eqn{n} is
the number of observations and \eqn{k} is the number of covariates
(including the intercept).}

\item{chain}{To be added}

\item{set}{Indicator for the use of set observations (1: set observations,
0: point observations).}

\item{eps_l}{Lower imprecision \eqn{(\epsilon_l)} for set observations
(default value: 0.5).}

\item{eps_r}{Upper imprecision \eqn{(\epsilon_r)} for set observations
(default value: 0.5)}
}
\description{
Deviance information criteria is based on the deviance funcion
\eqn{D(\theta, y) = -2 log(f(y|\theta))} but also incorporates a
penalization factor of the complexity of the model
}
