% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/User_Codes.R
\name{LML_LEP}
\alias{LML_LEP}
\title{Log-marginal likelihood estimator for the log-exponential power model}
\usage{
LML_LEP(
  thin,
  Time,
  Cens,
  X,
  chain,
  prior = 2,
  set = 1,
  eps_l = 0.5,
  eps_r = 0.5
)
}
\arguments{
\item{thin}{Thinning period.}

\item{Time}{Vector containing the survival times.}

\item{Cens}{Censoring indication (1: observed, 0: right-censored).}

\item{X}{Design matrix with dimensions \eqn{n} x  \eqn{k} where \eqn{n} is
the number of observations and \eqn{k} is the number of covariates
(including the intercept).}

\item{chain}{MCMC chains generated by a BASSLINE MCMC function}

\item{prior}{Indicator of prior (1: Jeffreys, 2: Type I Ind. Jeffreys,
3: Ind. Jeffreys).}

\item{set}{Indicator for the use of set observations (1: set observations,
0: point observations).}

\item{eps_l}{Lower imprecision \eqn{(\epsilon_l)} for set observations
(default value: 0.5).}

\item{eps_r}{Upper imprecision \eqn{(\epsilon_r)} for set observations
(default value: 0.5)}
}
\description{
Log-marginal likelihood estimator for the log-exponential power model
}
\examples{
library(BASSLINE)

# Please note: N=1000 is not enough to reach convergence.
# This is only an illustration. Run longer chains for more accurate
# estimations (especially for the log-exponential power model).

LEP <- MCMC_LEP(N = 1000, thin = 20, burn = 40, Time = cancer[,1],
                Cens = cancer[,2], X = cancer[,3:11])
LEP.LML <- LML_LEP(thin = 20, Time = cancer[,1], Cens = cancer[,2],
                   X = cancer[,3:11], chain = LEP)

}
