% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/User_Codes.R
\name{LML_LLOG}
\alias{LML_LLOG}
\title{Log-marginal likelihood estimator for the log-logistic model}
\usage{
LML_LLOG(thin, Q, Time, Cens, X, chain, prior, set, eps_l, eps_r, N.AKS = 3)
}
\arguments{
\item{thin}{Thinning period.}

\item{Q}{Update period for the \eqn{\lambda_{i}}’s}

\item{Time}{Vector containing the survival times.}

\item{Cens}{Censoring indication (1: observed, 0: right-censored).}

\item{X}{Design matrix with dimensions \eqn{n} x  \eqn{k} where \eqn{n} is
the number of observations and \eqn{k} is the number of covariates
(including the intercept).}

\item{chain}{MCMC chains generated by a BASSLINE MCMC function}

\item{prior}{Indicator of prior (1: Jeffreys, 2: Type I Ind. Jeffreys,
3: Ind. Jeffreys).}

\item{set}{Indicator for the use of set observations (1: set observations,
0: point observations).}

\item{eps_l}{Lower imprecision \eqn{(\epsilon_l)} for set observations
(default value: 0.5).}

\item{eps_r}{Upper imprecision \eqn{(\epsilon_r)} for set observations
(default value: 0.5)}

\item{N.AKS}{Maximum number of terms of the Kolmogorov-Smirnov density used
for the rejection sampling when updating mixing parameters (default
value: 3)}
}
\description{
Log-marginal likelihood estimator for the log-logistic model
}
