% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/User_Codes.R
\name{LML_LST}
\alias{LML_LST}
\title{Log-marginal Likelihood estimator for the log-student's t model}
\usage{
LML_LST(
  thin,
  Time,
  Cens,
  X,
  chain,
  Q = 1,
  prior = 2,
  set = 1,
  eps_l = 0.5,
  eps_r = 0.5
)
}
\arguments{
\item{thin}{Thinning period.}

\item{Time}{Vector containing the survival times.}

\item{Cens}{Censoring indication (1: observed, 0: right-censored).}

\item{X}{Design matrix with dimensions \eqn{n} x  \eqn{k} where \eqn{n} is
the number of observations and \eqn{k} is the number of covariates
(including the intercept).}

\item{chain}{MCMC chains generated by a BASSLINE MCMC function}

\item{Q}{Update period for the \eqn{\lambda_{i}}’s}

\item{prior}{Indicator of prior (1: Jeffreys, 2: Type I Ind. Jeffreys,
3: Ind. Jeffreys).}

\item{set}{Indicator for the use of set observations (1: set observations,
0: point observations). The former is strongly recommended over the
latter as point observations cause problems in the context of Bayesian
inference (due to continuous sampling models assigning zero probability
to a point).}

\item{eps_l}{Lower imprecision \eqn{(\epsilon_l)} for set observations
(default value: 0.5).}

\item{eps_r}{Upper imprecision \eqn{(\epsilon_r)} for set observations
(default value: 0.5)}
}
\description{
Log-marginal Likelihood estimator for the log-student's t model
}
\examples{
library(BASSLINE)

# Please note: N=1000 is not enough to reach convergence.
# This is only an illustration. Run longer chains for more accurate
# estimations.

LST <- MCMC_LST(N = 1000, thin = 20, burn = 40, Time = cancer[, 1],
                Cens = cancer[, 2], X = cancer[, 3:11])

LST.LML <- LML_LST(thin = 20, Time = cancer[, 1], Cens = cancer[, 2],
                   X = cancer[, 3:11], chain = LST)

}
