% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/User_Codes.R
\name{Trace_Plot}
\alias{Trace_Plot}
\title{Produce a trace plot of an observation's MCMC chain}
\usage{
Trace_Plot(obs = NULL, chain = NULL)
}
\arguments{
\item{obs}{Indicates the number of the observation under analysis}

\item{chain}{MCMC chains generated by a BASSLINE MCMC function}
}
\value{
A ggplot2 object
}
\description{
Plots the chain across (non-discarded) iterations for a
specified observation
}
\examples{
library(BASSLINE)

# Please note: N=1000 is not enough to reach convergence.
# This is only an illustration. Run longer chains for more accurate
# estimations.

LN <- MCMC_LN(N = 1000, thin = 20, burn = 40, Time = cancer[,1],
              Cens = cancer[,2], X = cancer[,3:11])
Trace_Plot(1, LN)

}
