% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/User_Codes.R
\name{BF_u_obs_LEP}
\alias{BF_u_obs_LEP}
\title{Outlier detection for observation for the log-exponential power model}
\usage{
BF_u_obs_LEP(
  N,
  thin,
  burn,
  ref,
  obs,
  Time,
  Cens,
  X,
  chain,
  prior,
  set,
  eps_l,
  eps_r,
  ar = 0.44
)
}
\arguments{
\item{N}{Total number of iterations.}

\item{thin}{Thinning period.}

\item{burn}{Burn-in period}

\item{ref}{Reference value \eqn{\lambda_{ref}} or \eqn{u_{ref}}}

\item{obs}{Indicates the number of the observation under analysis}

\item{Time}{Vector containing the survival times.}

\item{Cens}{Censoring indication (1: observed, 0: right-censored).}

\item{X}{Design matrix with dimensions \eqn{n} x  \eqn{k} where \eqn{n} is
the number of observations and \eqn{k} is the number of covariates
(including the intercept).}

\item{chain}{To be added}

\item{prior}{Indicator of prior (1: Jeffreys, 2: Type I Ind. Jeffreys,
3: Ind. Jeffreys).}

\item{set}{Indicator for the use of set observations (1: set observations,
0: point observations).}

\item{eps_l}{Lower imprecision \eqn{(\epsilon_l)} for set observations
(default value: 0.5).}

\item{eps_r}{Upper imprecision \eqn{(\epsilon_r)} for set observations
(default value: 0.5)}

\item{ar}{Optimal acceptance rate for the adaptive Metropolis-Hastings
updates}
}
\description{
This returns a unique number corresponding to the Bayes Factor
associated to the test \eqn{M_0: \Delta_{obs} = \lambda_{ref}} versus
\eqn{M_1: \Delta_{obs} \neq \lambda_{ref}} (with all other
\eqn{\Delta_j, \neq obs} free). The value of \eqn{\lambda_{ref}} is
required as input. The user should expect long running times for the
log-Student’s t model, in which case a reduced chain given
\eqn{\Delta_{obs} = \lambda_{ref}} needs to be generated
}
