% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/User_Codes.R
\name{LML_LEP}
\alias{LML_LEP}
\title{Log-marginal likelihood estimator for the log-exponential power model}
\usage{
LML_LEP(thin, Time, Cens, X, chain, prior, set, eps_l, eps_r)
}
\arguments{
\item{thin}{Thinning period.}

\item{Time}{Vector containing the survival times.}

\item{Cens}{Censoring indication (1: observed, 0: right-censored).}

\item{X}{Design matrix with dimensions \eqn{n} x  \eqn{k} where \eqn{n} is
the number of observations and \eqn{k} is the number of covariates
(including the intercept).}

\item{chain}{To be added}

\item{prior}{Indicator of prior (1: Jeffreys, 2: Type I Ind. Jeffreys,
3: Ind. Jeffreys).}

\item{set}{Indicator for the use of set observations (1: set observations,
0: point observations).}

\item{eps_l}{Lower imprecision \eqn{(\epsilon_l)} for set observations
(default value: 0.5).}

\item{eps_r}{Upper imprecision \eqn{(\epsilon_r)} for set observations
(default value: 0.5)}
}
\description{
Log-marginal likelihood estimator for the log-exponential power model
}
