% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/User_Codes.R
\name{MCMC_LEP}
\alias{MCMC_LEP}
\title{MCMC algorithm for the log-exponential power model}
\usage{
MCMC_LEP(
  N,
  thin,
  Time,
  Cens,
  X,
  beta0,
  sigma20,
  alpha0,
  prior,
  set,
  eps_l,
  eps_r,
  ar = 0.44
)
}
\arguments{
\item{N}{Total number of iterations.}

\item{thin}{Thinning period.}

\item{Time}{Vector containing the survival times.}

\item{Cens}{Censoring indication (1: observed, 0: right-censored).}

\item{X}{Design matrix with dimensions \eqn{n} x  \eqn{k} where \eqn{n} is
the number of observations and \eqn{k} is the number of covariates
(including the intercept).}

\item{beta0}{Starting value for \eqn{\beta}.}

\item{sigma20}{Starting value for \eqn{\sigma^2}.}

\item{alpha0}{Starting value for \eqn{\alpha}}

\item{prior}{Indicator of prior (1: Jeffreys, 2: Type I Ind. Jeffreys,
3: Ind. Jeffreys).}

\item{set}{Indicator for the use of set observations (1: set observations,
0: point observations).}

\item{eps_l}{Lower imprecision \eqn{(\epsilon_l)} for set observations
(default value: 0.5).}

\item{eps_r}{Upper imprecision \eqn{(\epsilon_r)} for set observations
(default value: 0.5)}

\item{ar}{Optimal acceptance rate for the adaptive Metropolis-Hastings
updates}
}
\description{
Markov chain Monte carlo algorithm for the log-exponential model
}
\examples{
library(SMLN)
n <- dim(cancer)[1]
# Building the Design Matrix
Intercept <- rep(1, times = n) # Intercept
x1 <- rep(0, times = n) # Treat (1 dummy variable)
for(i in 1 : n) {if(cancer$treat[i] == 2) x1[i] <- 1}
# Type (3 dummy variables)
x2 <- rep(0, times = n)
x3 <- rep(0, times = n)
x4 <- rep(0, times = n)

for(i in 1 : n) {if(cancer$type[i] == 1) x2[i] <- 1
if(cancer$type[i] == 2) x3[i] <- 1
if(cancer$type[i] == 3) x4[i] <- 1}
x5 <- cancer$status # Status
x6 <- cancer$mfd # Months from diagnosis
x7 <- cancer$age # Age
x8 <- rep(0, times = n) # Prior (Dummy variable)
for(i in 1 : n) {if(cancer$prior[i] == 10) x8[i] <- 1}
X <- cbind(Intercept, x1, x2, x3, x4, x5, x6, x7, x8)
Time <- cancer$time; Cens <- cancer$censor

# Set random starting values
beta0 <- rnorm(9, 0, 1)
sigma20 <- rgamma(1, 2, 2)
alpha0 <- runif(1, 1, 2)

# Please note: N=1000 is not enough to reach convergence.
# This is only an illustration. Run longer chains for more accurate
# estimations (especially for the log-exponential power model).

LEP <- MCMC_LEP(N = 1000, thin = 20, Time, Cens, X, beta0, sigma20,
                alpha0, prior = 2, set = 1, eps_l = 0.5, eps_r = 0.5,
                ar = 0.44)

}
