% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/User_Codes.R
\name{Trace_plot}
\alias{Trace_plot}
\title{Produce a trace plot of a variable's MCMC chain}
\usage{
Trace_plot(variable = NULL, chain = NULL)
}
\arguments{
\item{variable}{Indicates the index of the variable}

\item{chain}{MCMC chains generated by a BASSLINE MCMC function}
}
\value{
A ggplot2 object
}
\description{
Plots the chain across (non-discarded) iterations for a
specified observation
}
\examples{
library(BASSLINE)

# Please note: N=1000 is not enough to reach convergence.
# This is only an illustration. Run longer chains for more accurate
# estimations.

LN <- MCMC_LN(N = 1000, thin = 20, burn = 40, Time = cancer[, 1],
              Cens = cancer[, 2], X = cancer[, 3:11])
Trace_plot(1, LN)

}
